Commit #5 12/03/2020

+++ src/controllers/notes.controller.js
@@ -8,7 +8,10 @@ notesCtrl.renderNoteForm = (req, res) => {
 notesCtrl.createNewNote = async (req, res) => {
     const { title, description } = req.body;
     const newNote = new Note({ title, description });
+    newNote.user = req.user.id;
+
     await newNote.save();
+
 
     req.flash("success_msg", "Note added Successfully");
 
@@ -16,12 +19,18 @@ notesCtrl.createNewNote = async (req, res) => {
 };
 
 notesCtrl.renderNotes = async (req, res) => {
-    const notes = await Note.find();
+    const notes = await Note.find({ user: req.user.id }).sort({ createdAt: "desc" });
     res.render("notes/all-notes", { notes });
 }
 
 notesCtrl.renderEditForm = async (req, res) => {
     const note = await Note.findById(req.params.id);
+
+    if (note.user !== req.user.id) {
+        req.flash("error_msg", "Not authorizated.");
+        return res.redirect("/notes");
+    }
+
     res.render("notes/edit-note", { note });
 };
 
@@ -36,6 +45,13 @@ notesCtrl.updateNote = async (req, res) => {
 };
 
 notesCtrl.deleteNote = async (req, res) => {
+    const noteToDelete = Note.findById(req.params.id);
+
+    if (noteToDelete.user !== req.user.id) {
+        req.flash("error_msg", "Not authorizated.");
+        return res.redirect("/notes");
+    }
+
     await Note.findByIdAndDelete(req.params.id);
 
     req.flash("success_msg", "Note deleted Successfully");
